requirements.txt
aiogram==3.19.0
python-dotenv==1.1.0
yt_dlp==2025.3.31
